requirements.txt
Flask==2.2.2
requests==2.18.4
Jinja2==3.1.2
psycopg2-binary==2.9.5
